% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_scaling.R
\name{variable_features}
\alias{variable_features}
\title{Variable feature selection and scaling of seurat object}
\usage{
variable_features(input)
}
\arguments{
\item{input}{is the file after variable feature selection and scaling of the seurat object}
}
\description{
This package contains functions that identifies the variable features in a seurat object after quality filtering and normalization. The second function is scaling which should be done after variable feature identification.
}
\examples{
variable_features("input.rds")
}
